Add 'npm run profile' script to help profiling
