+++ .gitignore
@@ -9,3 +9,5 @@ node_modules/
 .vscode/
 *.generated.ts
 *.generated.spec.ts
+v8.log
+v8.out
